Move events dates functions to utils
